pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Oleg <igorkleschunov@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.4"
fastapi = "^0.75.1"
uvicorn = "^0.17.6"
gunicorn = "^20.1.0"
alembic = "^1.7.7"
psycopg2-binary = "^2.9.3"
invoke = "^1.7.0"
SQLAlchemy = "^1.4.35"
databases = {extras = ["postgresql"], version = "^0.5.5"}
click = "^8.1.2"
python-dotenv = "^0.20.0"
uvloop = "^0.16.0"
argon2-cffi = "^21.3.0"
itsdangerous = "^2.1.2"
structlog = "^21.5.0"
orjson = "^3.6.7"
nest-asyncio = "^1.5.5"
pydantic = {extras = ["email"], version = "^1.9.0"}
httpx = "^0.22.0"
asgi-lifespan = "^1.0.1"
PyJWT = "^2.3.0"
python-multipart = "^0.0.5"
pytest-asyncio = "^0.18.3"
Authlib = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
requests = "^2.27.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
